Refactor layout component path
Moved the layout component and related files to 'shared/components/layout' to better organize the project structure. Updated all import paths referencing the layout component accordingly.
