+++ frontend/src/app/shared/components/layout/layout.scss
